fix(ui): make file upload of game more obvious
- chore: Move ui components for game upload into new file
- fix(ui): add subtitle to file upload to make it more obvious that the
  icon does something, matching style of Logo upload component
