+++ public/locales/en/translation.json
@@ -89,6 +89,7 @@
   "Logo image is too large. 2MB is the limit": "Logo image is too large. 2MB is the limit",
   "Unknown file type": "Unknown file type",
   "logo upload": "logo upload",
+  "upload game": "upload game",
   "(must be smaller than 2MB)": "(must be smaller than 2MB)",
   "error reading file": "error reading file",
   "Unknown file type in game load": "Unknown file type in game load",
